fix(lint): updated type annotations for various to explicitly specify types

+++ packages/sealed_countries/test/src/test_data.dart
@@ -2,7 +2,8 @@ abstract final class TestData {
   const TestData._();
 
   static const integer = 1;
-  static const float = integer + 0.0;
+// ignore: avoid-explicit-type-declaration, vs specify_nonobvious_property_types.
+  static const double float = integer + 0.0;
 
   static const string = "$integer";
   static const emptyString = "";
